+++ examples/server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # server
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - phecda-server@1.0.2
+
 ## 0.0.2
 
 ### Patch Changes