+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # phecda-vue
 
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - phecda-core@1.0.8
+
 ## 1.2.1
 
 ### Patch Changes